requirements.txt
Flask==2.3.3
sentence-transformers==2.6.1
faiss-cpu==1.7.4
mysql-connector-python==8.3.0
python-dotenv==1.0.1
PyPDF2==3.0.1
loguru==0.7.2
PyYAML==6.0.1
